fix(scripts): fechar os defeitos da segunda ronda e o que eles revelaram
Num so commit porque os ficheiros se cruzam: separar as correcoes do refactor
deixava um commit intermedio com o CI vermelho (o check-doc-versions veria
citacoes a um anti-padrao que so existe no outro lado do par).

Reportados:
- a excecao do `if:` exigia exactamente um espaco a volta do `==`, logo
  `github.event_name=='pull_request'` contava como enfraquecimento e fechava o
  gate num ci.yml legitimo;
- o ramo de "ninguem cita" do Guard 15 era codigo morto: os cabecalhos das
  definicoes contavam como citacoes de si mesmos, logo `citacoes >= definicoes`
  sempre. A razao escrita no lugar do teste culpava este repo quando era de
  construcao, e valia em qualquer projeto;
- as definicoes ignoravam comentarios HTML e as citacoes nao, logo um projeto
  acabado de clonar levava aviso sobre uma linha dentro de `<!-- -->`.

Encontrados a corrigir os anteriores:
- a excecao do `if:` tambem nao cobria `${{ … }}` nem o escalar YAML entre
  aspas, e olhava so ao prefixo — `== 'pull_request' && false` era excluido.
  Passa a ancorar nos dois extremos;
- o strip de comentarios aplicado a `.mjs` cegava o guard entre dois literais
  `<!--`/`-->`, no proprio ficheiro que a correcao editou. Corre so em markdown;
- o strip encurtava o ficheiro e deslocava os numeros de linha citados nos
  avisos. Passa a preservar os `\n`;
- o teste do ramo NOTE herdava as definicoes do repo (AP3): a fixture passa a
  deriva-las do que o codigo dela cita;
- o padrao do cabecalho estava escrito duas vezes; alargar uma so devolvia o
  codigo morto em silencio. Passa a haver um literal;
- o ramo de "ninguem cita" continuava sem significado num projeto real, porque
  os comentarios dos proprios verificadores contam como citacoes. Passa a
  contar so documentacao; as citacoes em codigo continuam a valer para a
  deteccao de citacoes mortas, que e o que o guard protege.

Encontrados pela simulacao de um projeto derivado, que nunca tinha sido corrida:
- a checklist do BOOTSTRAP.md mandava deixar o anti-patterns.md "sem entradas".
  Nesse estado os ficheiros do template deixam 68 citacoes penduradas e o
  consumidor leva exit 1 no dia 1. A 2.8 e a nota do AP6 diziam o mesmo por
  outras palavras. O caminho suportado passa a ser manter as entradas e
  acrescentar com o proximo ID livre;
- o gate da superficie de teste comparava por ficheiro quando o invariante e
  sobre o TOTAL, e por isso reprovava as extracoes que o core-rules.md manda
  fazer acima das 500 linhas. Uma descida com o total intacto sai agora como
  NOTE ("movido, nao perdido");
- esvaziar as tabelas de padroes desligava a propria contagem que as vigiaria —
  exit 0 sobre o detetor desligado. Fechado com um invariante escrito a mao,
  fora dos dados.

Refactor forcado pelo flag das 500 linhas: Guard 15 para guards/anti-patterns.mjs,
os guards da serie 1 para guards/budgets.mjs, e o test-test-surface partido em
harness, marcas e tabelas de padroes. Nenhum ficheiro fica acima de 500 linhas e
a soma dos sitios de mutacao mantem-se.

O anti-patterns.md passa de 11934 para 9654 bytes (gate 12000): a evidencia vive
agora em src/docs/anti-patterns-why.md, nao carregado — o mesmo par que o
ticket-method ja usava.

Verificacao: check-doc-versions, check-backlog e check-test-surface a sair 0;
test-guards 172, test-test-surface 63, test-bundle-sizes 16, test-backlog 25,
test-mutation-sweep 17, test-hooks 156; mutation-sweep com cobertura completa
(14 alvos, 114 sitios). Os mesmos numeros num projeto derivado simulado.
