+++ .agent/rules/scripts-guide.md
@@ -11,6 +11,7 @@
 
 - **Doc Guards** (`.agent/scripts/check-doc-versions.mjs` + `guards/*.mjs`): Guards que correm sem config — orcamento de bytes das rules carregadas (NOTE 11 500 / gate 12 000), das de **referencia** (NOTE 12 500 / gate 14 000, com mensagem mais dura acima de 20 000 — limiares diferentes porque uma entra no contexto a cada sessao e a outra por ticket) e do **contexto** carregado (`.agent/context/`: NOTE 36 000 / gate 48 000, derivado das regras de arquivamento do `process-rules.md`; as rules ficam fora desta soma porque tem dono proprio no primeiro orcamento), paridade `CLAUDE.md`≡`GEMINI.md`, workflows↔wrappers (existencia **e** conteudo do ponteiro), workflows listados em `CLAUDE`/`GEMINI`/`AGENTS`/`agent-guide`, `@imports` resolvem, sanidade do `.claude/settings.json` (deny de secrets, allow sem wildcards abertos), **placeholders esquecidos apos o bootstrap** (salta enquanto o bootstrap nao correu), **as Fronteiras copiadas nos ponteiros do Cursor e do Copilot** (copia forcada: nao esta verificado que esses tools sigam um ponteiro em markdown, logo o bloco esta la e o guard compara-o), versao `package.json`≡`CHANGELOG`, `.nvmrc`, termos obsoletos — mais versoes de deps (configuravel). Caminhos ancorados a raiz do repo (nao ao cwd) e **todo o skip e visivel**: um guard que nao corre imprime `SKIP`, nunca desaparece. Sai `!= 0` em warning (serve de gate). **Corre no CI** no job `guard-tests` — e nao no `quality`, que salta sem `package.json` e deixava estes guards testados mas nunca aplicados. Correr tambem antes de commit e apos Dependabot PRs.
 - **Bundle Size Checker** (`.agent/scripts/check-bundle-sizes.mjs`): mede o First Load JS por rota contra os `TARGETS`. **Prefere falhar a reportar um numero que nao mediu** — ficheiro do manifest ausente do disco, ou rota cujos chunks proprios nao resolvem, saem `!= 0` em vez de imprimir `[OK]`. Limite conhecido no cabecalho: as rotas do App Router nao estao no `build-manifest.json`, e o varrimento de diretorio nao cobre route groups. Testes: `test-bundle-sizes.mjs`.
+- **Modulos de guard** (`.agent/scripts/guards/*.mjs`): guards extraidos do ficheiro de entrada quando ele passa o flag das 500 linhas — `budgets` (serie 1: tres orcamentos de bytes + a copia das Fronteiras), `settings` (11), `versions` (3 + deps), `derived-counts` (12/12c/12d/12e), `placeholders` (13) e `anti-patterns` (15: as citacoes de anti-padroes resolvem). Cada um leva a sua entrada em `PARES` e a sua suite `tests-*.mjs` espelhada; sem a entrada, a varredura mede o ficheiro de entrada e reporta 100% a mentir, porque os avisos passaram a viver no modulo. **O `test-test-surface.mjs` segue o mesmo desenho** desde que tambem passou o flag: harness em `test-surface-harness.mjs` (com o `resumo()`, porque os contadores sao de la) e o bloco das marcas em `tests-surface-marks.mjs`; as tabelas de padroes (dados, nao decisoes) em `surface-patterns.mjs`, que **nao** pode chamar-se `check-*` nem viver em `guards/` — a descoberta do sweep exigiria um par, e um par com zero sitios reprova com `SINAL ERRADO`. Esse ficheiro tem entrada propria em `CONFIG_CONTAVEIS` (senao esvaziar as tabelas saia da superficie vigiada) e um invariante **escrito a mao** no verificador a contar as entradas das tabelas: as CONTAGENS usam as tabelas actuais, logo esvazia-las desligava a propria contagem que as vigiava — exit 0 sobre o detetor desligado. **Uma extracao nao fecha o gate da superficie de teste**: o `check-test-surface.mjs` compara o TOTAL e nao so cada ficheiro, logo uma contagem que desce na origem e sobe no modulo novo sai como `NOTE` ("movido, nao perdido") e nao `WARN` — a descida do total continua a reprovar.
 - **Suites de testes negativos** (`test-guards.mjs`, `test-bundle-sizes.mjs`, `test-backlog.mjs`, `test-mutation-sweep.mjs`, `test-test-surface.mjs`, e `.claude/hooks/tests/test-hooks.mjs`): uma por verificador — quebram o que ele promete verificar e afirmam que avisa e sai `!= 0`. Correm no job `guard-tests` do CI, sem gate. Sem dependencias nem `package.json`. Correr apos qualquer alteracao a um `check-*.mjs`: um guard que passa quando devia falhar produz confianca infundada.
 - **Backlog Checker** (`.agent/scripts/check-backlog.mjs`): Trata os contadores do Resumo e a barra de progresso do backlog como dados derivados — recalcula-os a partir das tabelas (items abertos em `backlog.md` + fechados em `backlog-archive.md`) e avisa se divergirem; deteta tambem IDs duplicados. Caminhos ancorados a raiz do repo (nao ao cwd) e **ficheiro ausente reprova** — um gate que nao consegue validar nao pode sair `0`. Correr antes de commit; **corre no CI** no job `guard-tests`.
 - **Test Surface Checker** (`.agent/scripts/check-test-surface.mjs`): responde a "a superficie de teste foi enfraquecida desde a baseline?" — testes apagados, `skip`/`only` acrescentados, ou a **configuracao do runner** estreitada (que remove falhas sem tocar num teste). Compara **contagens** entre as duas versoes, nao linhas do diff. Sai `!= 0` tambem quando **nao consegue medir**. Ver `AP4`.
@@ -25,6 +26,14 @@
   - Ambos usam `git status --untracked-files=all`, porque sem isso o git **colapsa diretorios** nao rastreados e um ficheiro novo em pasta nova aparece como a pasta.
   - **Nao ha `lint-changed-file`** de proposito: os comandos de lint sao especificos da stack, logo o template so poderia trazer um hook inerte — e um hook que nao faz nada por omissao e prosa com mais passos. Se o teu projeto tem lint, vale a pena escreve-lo: `PostToolUse`/`Write|Edit`, a devolver o que nao e auto-corrigivel como contexto para ser corrigido no mesmo turno.
 - **Mutation Sweep** (`.agent/scripts/mutation-sweep.mjs`): mede se as suites **afirmam** algo — desliga cada sitio de erro de cada verificador, um a um, e exige que a suite fique vermelha. Sai `!= 0` se um sitio puder ser desligado com a suite verde, se um verificador nao tiver suite, ou se a baseline ja estiver vermelha. Custa minutos (recorre a suite por sitio), logo e opt-in no CI: correr localmente apos mexer num `check-*.mjs`. **Substitui contar sitios a mao** — o numero e derivado. **Varre-se a si proprio** (`--only=mutation-sweep`): reprova quem nao tem suite, logo nao pode ser a excecao.
+
+> **O limite da varredura, e o `AP7`.** Ela mede se cada sitio de aviso **existente** e
+> observado; **nao** mede se um ramo nunca dispara, nem ve um falso positivo. Um `note()` nao e
+> sitio de aviso, e um `warn()` inalcancavel por construcao passa igualmente — ela nao tem como
+> distinguir "coberto" de "impossivel". Medido: `26/26` no Guard 15 antes **e** depois de fechar
+> um ramo que era codigo morto. O que apanha isto e o **controlo negativo por ramo** — desligar
+> cada metade da correcao, uma por vez, e exigir vermelho em cada uma. Historia completa e as
+> licoes transferiveis em `src/docs/anti-patterns-why.md` (`AP7`).
 
 > **`test-harness.mjs`, `tests-*.mjs` e `.claude/hooks/tests/` nao sao entry points.** Correm
 > por importacao a partir do `test-guards.mjs` (ou do runner dos hooks) e reprovam se alguem os
